6.2
i figured out how to pass the direction of the raycast into the force applied to the ragdoll. all good there
also fixed some code. if you switched weapons before the reload finished and then equipped the weapon again, it would stay at 0 bullets. just added an else{startcoroutine reload()) to the shoot button
and moved the audioclip sound so it wouldnt play multiple times if you clicked shoot at 0 bullets
i put a random.range inside the y force of the direction passed to the ragdoll applyforce so sometimes it launches the ragdoll straight back, and sometimes it launches them into the air
i also had the colliders being destroyed when the zombie dies so that if the ragdoll hit the player they wouldnt take damage, but that was making them fall through the ground, so instead of it destroying all the colliders during the die function, i just made it change all the collider tags to "booty" instead of zombie so it wont trigger playerdamage